Moved the map marker constants into the utils directory
Cleaned up the structure in the models directory a little bit, now it only includes types
